// src/lib/api/schemas/role.ts
/**
 * Generated by orval v8.20.0 🍺
 * Do not edit manually.
 * ciclo
 * OpenAPI spec version: 0.1.0
 */
import type { RoleId } from './roleId.ts';
import type { RoleStatus } from './roleStatus.ts';

export interface Role {
  created_at: string;
  display_name: string;
  id: RoleId;
  slug: string;
  status: RoleStatus;
}